requirements.txt
sentence-transformers
pypdf
pinecone
python-dotenv
langchain-openai
langchain-pinecone
langchain-community
langchain-huggingface
langchain-experimental
-e .